Move SchemaManager into its own module - schema
